Move parser to the root of labelbox
To avoid loading of optional packages in data/ on python import
